move parsing logic into cmd/helm/template
